[Issue-690] refactor type
